added strong cloth and speed cloth

+++ 3D Adventure Game/Assets/Scripts/Interfaces/IDamageble.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public interface IDamageble
 {
-    public void Damage(int value);
+    public void Damage(float value);
 
-    public void Damage(int value, Vector3 dir);
+    public void Damage(float value, Vector3 dir);
 }
